backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
opencv-python==4.9.0.80
ultralytics==8.1.11
torch==2.1.2
torchvision==0.16.2
numpy==1.24.3
Pillow==10.2.0
shapely==2.0.2
paho-mqtt==1.6.1
pyyaml==6.0.1
python-dateutil==2.8.2
psutil==5.9.8
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
black==24.1.1
pylint==3.0.3
